Api.py userFolder changes

+++ snaw-backend/api.py
@@ -23,6 +23,7 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.secret_key = 'secret key'
 app.config['SESSION_TYPE'] = 'filesystem'
 DEBUG_FLAG = True
+count = 0
 '''
 ###------------------------------------------------------###
 App Routing: '/'
@@ -170,7 +171,8 @@ def run_analysis():
 
 @app.route('/removeUserFolder', methods=['GET', 'POST'])
 def closeUserFolder():
-    shutil.rmtree('instance/upload/user' + session['id'])
+    if os.path.isdir('instance/upload/user'+session['id']):
+        shutil.rmtree('instance/upload/user' + session['id'])
     return redirect('', 204)
